Refactored how we embed gstreamer.
Put the code that handles embedding gstreamer inside 1 platform-specific
class.  This is in preparation of overhauling the windows version.

I also tried to generally clean up the code.  We now clean up VideoWidget
properly, so we don't have to keep around a global singleton for it.
Simplified the Renderer interface.  Now all the renderer care's about is a
window id to pass it it's imagesink.  It doesn't need to connect to the
expose/destroy signals.